requirements.txt
# ALIS Backend Python Dependencies

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# LangGraph and LangChain
langgraph==0.2.0
langchain==0.3.0
langchain-core==0.3.0
openai==1.10.0 # Add OpenAI library

# Google Cloud & Firebase
google-cloud-firestore==2.14.0
firebase-admin==6.4.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Type Checking (optional, for development)
mypy==1.8.0
types-requests==2.31.0.10

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code Quality (optional)
black==23.12.1
flake8==7.0.0
isort==5.13.2
